+++ record_classification/src/main/java/uk/ac/standrews/cs/digitising_scotland/record_classification/classifiers/OLR/OLRShuffled.java
@@ -49,7 +49,7 @@ public class OLRShuffled implements Runnable {
 
     //----constructors---
 
-    public int getNumTrained(){
+    public long getNumTrained(){
         return model.getNumTrained();
     }
